requirements.txt
altgraph==0.10.2
appnope==0.1.0
audioread==2.1.2
awscli==1.8.2
backports.ssl-match-hostname==3.4.0.2
backports.weakref==1.0.post1
bdist-mpkg==0.5.0
bleach==1.5.0
bonjour-py==0.3
boto==2.39.0
botocore==1.2.0
bz2file==0.98
certifi==2015.9.6.2
click==6.6
colorama==0.3.3
config-reader==0.8
configobj==5.0.6
cycler==0.10.0
dask==0.10.0
decorator==4.0.10
docutils==0.12
enum==0.4.6
enum34==1.1.6
Flask==0.11.1
freenect==0.0.0
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
gensim==0.12.4
gnureadline==6.3.3
h5py==2.7.1
html5lib==0.9999999
httpretty==0.8.10
ipdb==0.10.3
ipykernel==4.0.3
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.0.2
itsdangerous==0.24
Jinja2==2.8
jmespath==0.7.1
joblib==0.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.0.0
jupyter-console==4.0.2
jupyter-core==4.0.4
Keras==2.0.8
leveldb==0.193
librosa==0.4.2
lxml==4.0.0
macholib==1.5.1
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==1.5.1
mercurial==3.7.3
mistune==0.7.1
mock==2.0.0
modulegraph==0.10.4
nbconvert==4.0.0
nbformat==4.0.0
networkx==1.11
notebook==4.0.4
numpy==1.13.1
olefile==0.44
path.py==8.1.1
pbr==3.1.1
pexpect==3.3
pickleshare==0.5
Pillow==4.2.1
protobuf==3.1.0
ptyprocess==0.5
py2app==0.7.3
pyasn1==0.1.8
pydot==1.0.2
Pygments==2.0.2
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.0.1
redis==2.10.5
requests==2.8.1
rsa==3.1.4
rudix==2015.6.1
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.3.2
stevedore==1.12.0
subprocess32==3.2.7
tensorflow==1.3.0
tensorflow-tensorboard==0.1.6
terminado==0.5
toolz==0.8.0
torch==0.1.11.post5
tornado==4.2.1
traitlets==4.0.0
update==0.4.4
vboxapi==1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
Werkzeug==0.12.2
xattr==0.6.4
yasm==0.0
zope.interface==4.1.1
